requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
Django==4.2.11
docarray==0.32.1
frozenlist==1.4.1
greenlet==3.0.3
hnswlib==0.8.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.43
langchain-text-splitters==0.0.1
langsmith==0.1.48
markdown-it-py==3.0.0
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.1
packaging==23.2
protobuf==5.26.1
pydantic==1.10.8
pydantic_core==2.18.1
Pygments==2.17.2
pypdf==4.2.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
tenacity==8.2.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.4
Django==4.2.11
djangorestframework==3.15.1
docarray==0.32.1
frozenlist==1.4.1
greenlet==3.0.3
hnswlib==0.8.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.43
langchain-text-splitters==0.0.1
langsmith==0.1.48
markdown-it-py==3.0.0
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.1
packaging==23.2
protobuf==5.26.1
pydantic==1.10.8
pydantic_core==2.18.1
Pygments==2.17.2
pypdf==4.2.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
tenacity==8.2.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
